requirements.txt
# Required libraries to be installed
split-folders
omrdatasettools
torchvision
tqdm
pillow
torch
matplotlib
tensorflow
keras
Keras-Preprocessing